refactored Instacne components
